+++ crates/primitives/block/src/lib.rs
@@ -3,7 +3,6 @@
 
 #[doc(hidden)]
 pub extern crate alloc;
-
 use alloc::vec::Vec;
 
 mod header;
@@ -29,6 +28,7 @@ pub type BlockEvents = Vec<OrderedEvents>;
 #[derive(Debug, Clone, Copy, PartialEq, Eq)]
 #[cfg_attr(feature = "parity-scale-codec", derive(parity_scale_codec::Encode, parity_scale_codec::Decode))]
 #[cfg_attr(feature = "serde", derive(serde::Deserialize, serde::Serialize))]
+#[cfg_attr(feature = "scale-info", derive(scale_info::TypeInfo))]
 pub enum BlockTag {
     #[cfg_attr(feature = "serde", serde(rename = "latest"))]
     Latest,